Fix handling of {log_error, Code, Req, Reason} events
Although the log_error/3 event in webmachine_log has a spec that
indicates that Req is a `#wm_reqdata{}` record, the webmachine code
actually sends `{webmachine_request, #wm_reqstate{}}`.

This patch adds helper functions wm_method, wm_path, and wm_notes to
deal with the expected form. These helpers return undefined if
unexpected data is received to avoid crashing the logger and also
handle receiving a #wm_reqdata{} record.

Also adjusted is the handling of log messages and Reason to take
advantage of the pre-existing handling of notes and msg
annotation. These messages are now merged into the msg annotation
field of notes and emitted as raw.

Tests adjusted for what we see from wm and an additional "bogus" test
added to verify non-crash behavior.

Finally, here are some notes on the log event flow in wm:

Looking at calls to gen_event:call and gen_event:sync_notify and the
specs for the calling functions in webmachine_log, I see the following [2]:

    %% @doc Notify registered log event handler of an error event.
    -spec log_error(pos_integer(), #wm_reqdata{}, term()) -> ok.
    log_error(Code, Req, Reason) ->
        gen_event:sync_notify(?EVENT_LOGGER, {log_error, Code, Req, Reason}).

However, it seems that the event logger will get called with a Req value
that has the shape {webmachine_request, #wm_reqstate{}}.

For example in webmachine_mochiweb:

    Req = webmachine:new_request(mochiweb, MochiReq)
    ...
    handle_error(500, {error, Error}, Req)

    https://github.com/basho/webmachine/blob/develop/src/webmachine_mochiweb.erl#L49
    https://github.com/basho/webmachine/blob/develop/src/webmachine_mochiweb.erl#L77

    where handle_error does:

        ErrorHandler:render_error(Code, Req, Error)

    and in default config that will be a call to
    webmachine_error_handler:render_error/3 which sure looks like it is
    expecting the tuple not #wm_reqdata{}:

    https://github.com/basho/webmachine/blob/develop/src/webmachine_error_handler.erl#L32
    https://github.com/basho/webmachine/blob/develop/src/webmachine_error_handler.erl#L93