Adjust format to capture most error traces on a single line
NOTE: It seems that we end up logging both the access and the
error. Not clear how to obtain the req id in the error case which
would at lest allow these to be aligned.

tempted to not log to request log in the error case, but to include
the request log format and the full Reason to console since that makes
it easier to line things up there.